requirements.txt
python-dotenv==0.13.0
spotipy==2.12.0
